requirements.txt
# The requirements listed here are for development. 
# Requirements of your napari plugin should be listed in setup.cfg
# See also: https://caremad.io/posts/2013/07/setup-vs-requirement/

-e .

pip~=21.3
wheel~=0.37.0
openssl~=3.0.0
tornado~=6.1
numpy~=1.20.3
setuptools~=59.2.0
docutils~=0.17.1
pandas~=1.3.3
napari~=0.4.11
matplotlib~=3.5.0
magicgui~=0.3.0
pyqt5~=5.15.4
qtpy~=1.11.2
scikit-image~=0.19
umap-learn~=0.5.1
scikit-learn~=1.0